refactored ui out of display
