+++ extension/content_script.js
@@ -1,6 +1,9 @@
 // Prevent the usage of undeclared variables.
 "use strict";
 
+// Is this used anywhere?
+// const { UI } = require("winjs");
+
 (function(){
 var namespace = "content_script.js";
 if (window[namespace] === true) {
@@ -13,6 +16,8 @@ if (window[namespace] === true) {
 var tracker = null;
 // Creates display for time remaining and reading speed at top of page. Initialized by end of script load.
 var display = null;
+// Creates UI for users to see instructions & customize settings
+var ui = null;
 // Represents document the user is reading. Stores data about page, like keywords, total words, etc.
 var doc = null;
 // Whether or not there is a timer that triggers movement of tracker.
@@ -29,7 +34,7 @@ var speed_bias = 500; // Minimum amount of speed spent on each sentence (in mill
 var isScrolling = false;
 
 // Need the same $ reference for on and off of event handlers to be detectable.
-// Re-doing $(document) in an async context for somre reason doesn't allow you 
+// Re-doing $(document) in an async context for some reason doesn't allow you 
 // to detect previously attached event handlers.
 // Not sure why, but found this through experimentation.
 let jdoc = $(document);
@@ -188,7 +193,7 @@ function scrollUp() {
 	if ($("."+trackerStyle)) {
 		tracker_style_current = trackerStyle
 	} else { 
-		let displaySettings = display.getSettings();
+		let displaySettings = ui.getSettings();
 		tracker_style_current = "trackerHighlighter"+displaySettings[1]+"Shadow"+displaySettings[2]; 
 	}
 
@@ -217,7 +222,7 @@ function scrollDown() {
 	if ($("."+trackerStyle)) {
 		tracker_style_current = trackerStyle
 	} else { 
-		let displaySettings = display.getSettings();
+		let displaySettings = ui.getSettings();
 		tracker_style_current = "trackerHighlighter"+displaySettings[1]+"Shadow"+displaySettings[2]; 
 	}
 	
@@ -291,7 +296,7 @@ function highlight(tracker) {
     */
 function highlightKeyWords(container, start, end) {
 	// TODO: Refactor this; this should be reset whenever it is changed, not checked every sentence
-	let displaySettings = display.getSettings();
+	let displaySettings = ui.getSettings();
 	let keywordStyle = "keyWord"+displaySettings[0]; 
 	// TODO: Mark.js is actually built to do this; migrate functionality to mark.js
 	$("."+keywordStyle).unmark(); // Remove previous sentence keyword styling
@@ -323,7 +328,7 @@ function highlightKeyWords(container, start, end) {
 Fade the current tracker indicator according to the calculated speed.
 */
 function fadeTracker() {
-	let displaySettings = display.getSettings();
+	let displaySettings = ui.getSettings();
 	let keywordStyle = "keyWord"+displaySettings[0];
 	fadeElement($("mark"));
 	fadeElement($("." + keywordStyle));
@@ -433,7 +438,6 @@ function initializeTracker(settingsCustomizations) {
 REMOVE THIS
 */
 function updateDisplaySettings() {
-	let displaySettings = []; // Settings: 1: Keyword 2: Highlighter 3: Shadow
 	settings.getCustomizations(function(settingsCustomizations) {
 		initializeTracker(settingsCustomizations)
 	});
@@ -486,6 +490,7 @@ Render all the UI elements.
 */
 function setupUI() {
 	display = new Display(doc, speed);
+	ui = new Ui();
   
 	updateDisplaySettings();
   
@@ -505,6 +510,7 @@ function removeUI() {
 	tracker.reset();
 	display.turnDownUI();
 	display = null;
+	ui = null;
 }
 
 function toggleExtensionVisibility() {
